edit code

+++ db.js
@@ -56,7 +56,6 @@ window.onload = function() {
                                 '<div class="dimmer"> <h6 class="h6">' + doc.data().timepostShow + '</h6></div>' +
                                 '<div class="dimmer"> <h6 class="h6">' + doc.data().timenotification + '</h6></div>' +
 
-                                '<button type="button" class="btn btn-outline-success border border-warning" style="width:100px; margin:30px 88px"  >More</button>' +
                                 '</div>' +
 
                                 '</a>' +
@@ -96,34 +95,34 @@ window.onload = function() {
                             storageRef.child('NoteImage/' + doc.data().image + '').getDownloadURL().then(function(url) {
                                 imgNote.src = url;
                             }).catch(function(error) {});
-                            if(linhDay >= curDay ){
+                            if (linhDay >= curDay) {
                                 $("#pills-doing").append(
-                                    '<div class="Toaster">'+
-                                            '<div class="Toaster-iconContainer">'+
-                                                '<div class="Toaster-icon Toaster-icon--success">'+
-                                                    '<i class="fa fa-check"></i>'+
-                                                '</div>'+
-                                            '</div>'+
-                                            '<div class="Toaster-messageContainer">'+
-                                                '<div class="Toaster-message">'+doc.data().name+'</div>'+
-                                            '</div>'+
-                                        '</div>'
+                                    '<div class="Toaster">' +
+                                    '<div class="Toaster-iconContainer">' +
+                                    '<div class="Toaster-icon Toaster-icon--success">' +
+                                    '<i class="fa fa-check"></i>' +
+                                    '</div>' +
+                                    '</div>' +
+                                    '<div class="Toaster-messageContainer">' +
+                                    '<div class="Toaster-message">' + doc.data().name + '</div>' +
+                                    '</div>' +
+                                    '</div>'
                                 );
-                            }else if(linhDay < curDay ){
+                            } else if (linhDay < curDay) {
                                 $("#pills-done").append(
-                                    '<div class="Toaster">'+
-                                            '<div class="Toaster-iconContainer">'+
-                                                '<div class="Toaster-icon Toaster-icon--error">'+
-                                                    '<i class="fa fa-exclamation"></i>'+
-                                                '</div>'+
-                                            '</div>'+
-                                            '<div class="Toaster-messageContainer">'+
-                                                '<div class="Toaster-message">'+doc.data().name+'</div>'+
-                                            '</div>'+
-                                        '</div>'
+                                    '<div class="Toaster">' +
+                                    '<div class="Toaster-iconContainer">' +
+                                    '<div class="Toaster-icon Toaster-icon--error">' +
+                                    '<i class="fa fa-exclamation"></i>' +
+                                    '</div>' +
+                                    '</div>' +
+                                    '<div class="Toaster-messageContainer">' +
+                                    '<div class="Toaster-message">' + doc.data().name + '</div>' +
+                                    '</div>' +
+                                    '</div>'
                                 );
                             }
-                        }  
+                        }
                     });
 
                 });
@@ -186,10 +185,7 @@ function Category_select(category) {
                                         '</div>' +
                                         '<p class="p small ">Content:' + subDes + '</p>' +
                                         '<div class="dimmer"> <h6 class="h6">' + doc.data().timepostShow + '</h6></div>' +
-
-                                        '<button type="button" class="btn btn-outline-success border border-warning" style="width:100px; margin:30px 88px"  >More</button>' +
                                         '</div>' +
-
                                         '</a>' +
 
                                         '<div class="modal fade" id="exampleModalScrollable' + doc.id + '" tabindex="-1" role="dialog" aria-labelledby="exampleModalScrollableTitle" aria-hidden="true">' +
@@ -256,6 +252,7 @@ function addnewcategory() {
             userID: idUser
         })
         alert("Add to category");
+        category_Name.value = '';
     }
 }
 
@@ -267,21 +264,12 @@ function deleteCategory() {
         alert('Comments are required to continue!');
 
     } else {
-        db.collection("category").doc(document.getElementById("category_Name").value + firebase.auth().currentUser.uid).delete();
+        if (confirm('You want to delete category?')) {
+            db.collection("category").doc(document.getElementById("category_Name").value + firebase.auth().currentUser.uid).delete();
+        } else {}
+        category_Name.value = '';
     }
 }
-
-// function deleteCategoryInSelect() {
-//   // var result = confirm("Do you want to continue?");
-//   // if(result)  {
-//   // alert("OK Next lesson!");
-//   db.collection("category").doc(document.getElementById("ListNotes_category1").value + firebase.auth().currentUser.uid).delete();
-//   // location.reload();
-//   // } else {
-//   // alert("Bye!");
-//   // }
-
-// }
 
 function checkTime(i) {
     if (i < 10) {
@@ -296,7 +284,6 @@ document.getElementById('NewNotes_checkTime').onclick = function(e) {
     } else {
         document.getElementById('notification').style.display = 'none';
     }
-
 };
 
 function save(id) {
@@ -380,7 +367,7 @@ function writeNotesData() {
     m2 = checkTime(m2);
     timepost2 = MM2 + "/" + DD2 + "/" + YYYY2 + " " + h2 + ":" + m2;
 
-    if(timepost2.includes("NaN")){
+    if (timepost2.includes("NaN")) {
         timepost2 = "";
     }
     alert(timepost2);
@@ -519,7 +506,7 @@ function search() {
     $('#listNotes').load('show_not.html');
     var tim = document.getElementById("search").value;
     console.log(tim);
-    firebase.auth().onAuthStateChanged(function (user) {
+    firebase.auth().onAuthStateChanged(function(user) {
         if (user) {
             var user = firebase.auth().currentUser;
 
@@ -528,7 +515,7 @@ function search() {
                 // var listcategory = document.getElementById("ListNotes_category").value;
                 db.collection("notelist").orderBy("timepost", "desc").get().then(snapshot => {
                     snapshot.docs.forEach(doc => {
-                        var tim2 = ""+(doc.data().name).toLowerCase();
+                        var tim2 = "" + (doc.data().name).toLowerCase();
                         if (doc.data().userID == idUser && tim2.includes(tim.toLowerCase())) {
                             var des = doc.data().content;
                             var tit = doc.data().name;
@@ -553,8 +540,6 @@ function search() {
                                 '</div>' +
                                 '<p class="p small ">Content:' + subDes + '</p>' +
                                 '<div class="dimmer"> <h6 class="h6">' + doc.data().timepostShow + '</h6></div>' +
-
-                                '<button type="button" class="btn btn-outline-success border border-warning" style="width:100px; margin:30px 88px"  >More</button>' +
                                 '</div>' +
 
                                 '</a>' +
@@ -589,11 +574,10 @@ function search() {
 
 
                             var imgNote = document.getElementById("imgNote" + doc.data().name);
-                            storageRef.child('NoteImage/' + doc.data().image + '').getDownloadURL().then(function (url) {
+                            storageRef.child('NoteImage/' + doc.data().image + '').getDownloadURL().then(function(url) {
                                 imgNote.src = url;
-                            }).catch(function (error) { });
-                        }
-                        else if (doc.data().userID == idUser && document.getElementById("search").value == "" && category_selected == "") {
+                            }).catch(function(error) {});
+                        } else if (doc.data().userID == idUser && document.getElementById("search").value == "" && category_selected == "") {
                             var des = doc.data().content;
                             var tit = doc.data().name;
                             var subTit = tit.slice(0, 30);
@@ -617,8 +601,6 @@ function search() {
                                 '</div>' +
                                 '<p class="p small ">Content:' + subDes + '</p>' +
                                 '<div class="dimmer"> <h6 class="h6">' + doc.data().timepostShow + '</h6></div>' +
-
-                                '<button type="button" class="btn btn-outline-success border border-warning" style="width:100px; margin:30px 88px"  >More</button>' +
                                 '</div>' +
 
                                 '</a>' +
@@ -653,9 +635,9 @@ function search() {
 
 
                             var imgNote = document.getElementById("imgNote" + doc.data().name);
-                            storageRef.child('NoteImage/' + doc.data().image + '').getDownloadURL().then(function (url) {
+                            storageRef.child('NoteImage/' + doc.data().image + '').getDownloadURL().then(function(url) {
                                 imgNote.src = url;
-                            }).catch(function (error) { });
+                            }).catch(function(error) {});
                         } else if (doc.data().userID == idUser && document.getElementById("search").value == "" && category_selected == doc.data().category) {
                             var des = doc.data().content;
                             var tit = doc.data().name;
@@ -680,8 +662,6 @@ function search() {
                                 '</div>' +
                                 '<p class="p small ">Content:' + subDes + '</p>' +
                                 '<div class="dimmer"> <h6 class="h6">' + doc.data().timepostShow + '</h6></div>' +
-
-                                '<button type="button" class="btn btn-outline-success border border-warning" style="width:100px; margin:30px 88px"  >More</button>' +
                                 '</div>' +
 
                                 '</a>' +
@@ -716,9 +696,9 @@ function search() {
 
 
                             var imgNote = document.getElementById("imgNote" + doc.data().name);
-                            storageRef.child('NoteImage/' + doc.data().image + '').getDownloadURL().then(function (url) {
+                            storageRef.child('NoteImage/' + doc.data().image + '').getDownloadURL().then(function(url) {
                                 imgNote.src = url;
-                            }).catch(function (error) { });
+                            }).catch(function(error) {});
                         }
                     });
                 });
